Options: Add --pattern-length-limit
Since 7b47ff2 (main: omit patterns for long input lines, 2015-10-15) long
tag lines are cut off after 96 characters.

For applications reading the tag files this can be unfortunate as it
breaks backwards compatibility with excuberant tags.

Introduce an option --patttern-length-limit=N which allows the user to set
an arbitrary cutoff value. The value 0 is special as it turns the length
limiting off.

Signed-off-by: Thomas Braun <thomas.braun@virtuell-zuhause.de>